+++ src/test/java/com/shyoon/wms/inbound/feature/ConfirmInboundTest.java
@@ -34,18 +34,4 @@ class ConfirmInboundTest {
         assertThat(inbound.getStatus()).isEqualTo(InboundStatus.CONFIRMED);
     }
 
-    private class ConfirmInbound {
-
-        private final InboundRepository inboundRepository;
-
-        private ConfirmInbound(InboundRepository inboundRepository) {
-            this.inboundRepository = inboundRepository;
-        }
-
-        public void request(final Long inboundNo) {
-            final Inbound inbound = inboundRepository.getBy(inboundNo);
-
-            inbound.confirmed();
-        }
-    }
 }
